Introducing the changes to split the appeals phase in two

+++ web_module/branches/OR_20041217/src/main/com/topcoder/apps/review/AppealAction.java
@@ -5,7 +5,7 @@
 package com.topcoder.apps.review;
 
 import com.topcoder.apps.review.document.Appeal;
-import com.topcoder.apps.review.Phase;
+import com.topcoder.apps.review.projecttracker.Phase;
 import com.topcoder.util.log.Level;
 
 import javax.servlet.http.HttpServletRequest;
@@ -95,7 +95,7 @@ public final class AppealAction extends ReviewAction {
                 request.setAttribute("reviewerEdit", new Boolean(true));
             }
 */
-            int phaseId = orpd.getProject().getCurrentPhaseInstance().getId();
+            long phaseId = orpd.getProject().getCurrentPhaseInstance().getId();
             if (appeal.getAppealer().getId() == orpd.getUser().getId() &&
                     appeal.getId() == -1 && phaseId == Phase.ID_APPEALS) {
                 request.setAttribute("appealerEdit", new Boolean(true));
